vulfocus-api/requirements.txt
Django==2.2.24
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
Markdown==3.1.1
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
IPy==1.0
docker==3.7.3
docker-compose==1.25.4
docker-pycreds==0.4.0
dockerpty==0.4.1
celery==4.4.2
redis==3.2.0
Pillow==8.2.0
dnspython==2.1.0
ruamel.yaml==0.17.10
django-simple-captcha==0.5.14
django-notifications-hq==1.6.0
zipfile36==0.1.3
PyMySQL==1.0.2
requests